Run the real fine-tune: LoRA training, GGUF quantization, Ollama serving, eval
Actually executed the pipeline end to end on Qwen2.5-Coder-0.5B-Instruct:
LoRA fine-tune converging cleanly over 3 epochs (eval loss 0.060 -> 0.008),
merged and quantized to a 373MB q4_k_m GGUF, served through Ollama, and
scored at 92.9% execution accuracy (26/28) with ~200ms latency and $0 cost.
The MCP server was verified end-to-end against the real served model.

Replaced transformers.Trainer with a ~40-line manual training loop after
isolating two real PyTorch memory bugs hit while actually running this on
Apple Silicon: an MPS caching-allocator runaway (23GB RSS, traced to
dynamic-shape padding) and unexplained Trainer/DataLoader idle time despite
a fast underlying forward/backward. Fixed-length batch padding eliminates
both failure classes rather than working around them.

Frontier-model baseline still needs a live ANTHROPIC_API_KEY to run --
eval/baseline_frontier.py is written and tested but not yet executed.

+++ serving/hf_predictor.py
@@ -18,7 +18,7 @@ class HFLocalPredictor:
         self.tokenizer = AutoTokenizer.from_pretrained(model_dir)
         device = "cuda" if torch.cuda.is_available() else ("mps" if torch.backends.mps.is_available() else "cpu")
         dtype = torch.bfloat16 if device == "cuda" else torch.float32
-        self.model = AutoModelForCausalLM.from_pretrained(model_dir, torch_dtype=dtype).to(device)
+        self.model = AutoModelForCausalLM.from_pretrained(model_dir, dtype=dtype).to(device)
         self.device = device
 
     def predict(self, question: str) -> Prediction:
